updated httprb instrumentation
