+++ test/multiverse/suites/httprb/Envfile
@@ -2,6 +2,8 @@ suite_condition("http.rb is only supported for versions >= 2.0.0") do
   RUBY_VERSION >= '2.0.0'
 end
 
+instrumentation_methods :chain, :prepend
+
 gemfile <<-RB
   gem 'http' # latest version
   gem 'public_suffix', '< 3.0.0' if RUBY_VERSION < "2.1.0"
