+++ test/multiverse/suites/httprb/config/newrelic.yml
@@ -5,6 +5,8 @@ development:
   apdex_t: 0.5
   monitor_mode: true
   license_key: bootstrap_newrelic_admin_license_key_000
+  instrumentation:
+    http: <%= $instrumentation_method %>
   ca_bundle_path: ../../../config/test.cert.crt
   app_name: test
   host: localhost
